题目相关接口开发

+++ src/main/java/com/mincai/ikuncode/model/dto/user/UserUpdateRequest.java
@@ -12,14 +12,20 @@ import java.io.Serializable;
 @Data
 public class UserUpdateRequest implements Serializable {
 
-    // 用户 id
+    /**
+     * 用户 id
+     */
     private Long userId;
 
-    // 用户昵称
+    /**
+     * 用户昵称
+     */
     private String userNickname;
 
-    // 用户密码
-    private Long userRole = -1L;
+    /**
+     * 用户角色
+     */
+    private Integer userRole = -1;
 
     private static final long serialVersionUID = 1L;
 }
